fix database

+++ src/main/java/com/ordering_be/ordering_be_spring/Models/NewMenuRequest.java
@@ -2,12 +2,14 @@ package com.ordering_be.ordering_be_spring.Models;
 
 import lombok.NonNull;
 
+import java.math.BigDecimal;
+
 public class NewMenuRequest {
 
     public String name;
 
     public String description;
 
-    public float price;
+    public BigDecimal price;
 
 }
